backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pdfplumber
PyPDF2
transformers
sentence-transformers
faiss-cpu
numpy
pandas
aiofiles
requests
python-dotenv
